# services/organization-service/openapi.yaml
openapi: 3.0.3
info:
  title: Organization Service
  version: 1.0.0
  description: Phase 1 tenant administration. Production product access is fail-closed
    pending verified Identity integration. All operations require the documented permission
    and trusted tenant context. Lists use cursor pagination.
paths:
...
  /api/v1/organizations/current/lifecycle/:
    post:
      operationId: v1_organizations_current_lifecycle_create
      description: |2-

        Required permission: `organization.update`. Tenant-scoped identifiers return 404 outside the current organization. Errors use `error.code`, `error.message`, `error.details`, `error.correlation_id`.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleInputRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LifecycleInputRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LifecycleInputRequest'
        required: true
      security:
      - DevelopmentOnly: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
          description: ''
  /api/v1/organizations/current/settings/:
    get:
      operationId: v1_organizations_current_settings_retrieve
      description: |2-

        Required permission: `organization.view`. Tenant-scoped identifiers return 404 outside the current organization. Errors use `error.code`, `error.message`, `error.details`, `error.correlation_id`.
      tags:
      - v1
      security:
      - DevelopmentOnly: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
          description: ''
    patch:
      operationId: v1_organizations_current_settings_partial_update
      description: |2-

        Required permission: `organization.update`. Tenant-scoped identifiers return 404 outside the current organization. Errors use `error.code`, `error.message`, `error.details`, `error.correlation_id`.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedSettingsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedSettingsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedSettingsRequest'
      security:
      - DevelopmentOnly: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
          description: ''
...
components:
  schemas:
...
    LifecycleInputRequest:
      type: object
      properties:
        status:
          type: string
          minLength: 1
          maxLength: 20
      required:
      - status
...
    Organization:
      type: object
      description: Reject unknown/read-only input rather than silently accepting tenant/identity
        spoofing.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        name:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 100
          pattern: ^[-a-zA-Z0-9_]+$
        display_name:
          type: string
          maxLength: 200
        organization_type:
          $ref: '#/components/schemas/OrganizationTypeEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/OrganizationStatusEnum'
          readOnly: true
        logo_media_id:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          format: email
          nullable: true
          maxLength: 254
        phone:
          type: string
          nullable: true
          maxLength: 32
        timezone:
          type: string
          maxLength: 64
        country_code:
          type: string
          maxLength: 2
        suspended_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        deleted_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - created_at
      - deleted_at
      - display_name
      - id
      - name
      - slug
      - status
      - suspended_at
      - updated_at
    OrganizationStatusEnum:
      enum:
      - ACTIVE
      - SUSPENDED
      - DISABLED
      - DELETED
      type: string
      description: |-
        * `ACTIVE` - Active
        * `SUSPENDED` - Suspended
        * `DISABLED` - Disabled
        * `DELETED` - Deleted
    OrganizationTypeEnum:
      enum:
      - BUSINESS
      - NONPROFIT
      - OTHER
      type: string
      description: |-
        * `BUSINESS` - Business
        * `NONPROFIT` - Nonprofit
        * `OTHER` - Other
...
    PatchedSettingsRequest:
      type: object
      description: Reject unknown/read-only input rather than silently accepting tenant/identity
        spoofing.
      properties:
        allow_member_invites:
          type: boolean
        allow_group_creation:
          type: boolean
        allow_external_chat:
          type: boolean
        max_members:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        max_teams:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        default_timezone:
          type: string
          minLength: 1
          maxLength: 64
...
    Settings:
      type: object
      description: Reject unknown/read-only input rather than silently accepting tenant/identity
        spoofing.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        allow_member_invites:
          type: boolean
        allow_group_creation:
          type: boolean
        allow_external_chat:
          type: boolean
        max_members:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        max_teams:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        default_timezone:
          type: string
          maxLength: 64
        organization:
          type: string
          format: uuid
          readOnly: true
      required:
      - created_at
      - id
      - organization
      - updated_at
...
  securitySchemes:
    DevelopmentOnly:
      type: apiKey
      in: header
      name: X-Dev-Member-ID
      description: Unsafe local-only identity injection. Disabled in production. Uses
        an ACTIVE database member and current database permissions.